fix(button): fix rerender issue
affects: @fremtind/jkl-button-react

defining button inside button caused an rerender all the way out to the DOM for each change in the
react tree

ISSUES CLOSED:  #1783
